requirements.txt
Flask
requests
python-dotenv
gunicorn
beautifulsoup4
lxml
selenium
tldextract
idna